Hides real implementations
