+++ pillarbox-analytics/src/androidTest/java/ch/srgssr/pillarbox/analytics/UserAnalyticsTest.kt
@@ -5,7 +5,7 @@
 package ch.srgssr.pillarbox.analytics
 
 import androidx.test.platform.app.InstrumentationRegistry
-import ch.srgssr.pillarbox.analytics.commandersact.CommandersAct
+import ch.srgssr.pillarbox.analytics.commandersact.CommandersActConfig
 import org.junit.Assert
 import org.junit.Test
 
@@ -13,7 +13,7 @@ class UserAnalyticsTest {
 
     private val config = SRGAnalytics.Config(
         analyticsConfig = AnalyticsConfig(distributor = AnalyticsConfig.BuDistributor.SRG),
-        commandersAct = CommandersAct.Config(virtualSite = "pillarbox-test-android", sourceKey = CommandersAct.Config.SOURCE_KEY_SRG_DEBUG)
+        commandersAct = CommandersActConfig(virtualSite = "pillarbox-test-android", sourceKey = CommandersActConfig.SOURCE_KEY_SRG_DEBUG)
     )
 
     @Test
